First implementation of history in the shell.
Changed cin structure to getch structure. That way it will interpret
each character as it is received. This allows me to grab arrow
characters.

Added a cursor_position counter that keeps track of the x position
relative to the prompt. This allows me to backspace using "\b \b".

Added history grabbing when you press up or down on the keyboard.