First implementation of external commands.
Added colors to prompt.

Removed InterpretCommand since it was cleaner to deal with the return
values etc in main.

A call to ExecuteExternalCommands deals calls execvp. execvp requires a
list of char pointers contiguous in memory. It is required that the last
pointer be equal to NULL, otherwise the data will get corrupted.

If execvp returns -1, there was an error, and the child process has to
deal with it.