Added job control and kill implementation.
Added jobs and kill command to builtin.h.
Also added implementations in builtin functions.cpp.

These basically toggle flags so that the corresponding functions are
called in main().

Added CheckJobs, which iterates through the job list and checks the
status of each one. Then it walks through the list and deletes
terminated and exited jobs.

Added Kill, which parses the %[jobnum] argument, and moves an iterator
to the correct spot and kills the corresponding process.